Track .docx with Git LFS

+++ src/main/java/com/EIPplatform/model/dto/report/airemmissionmanagement/airautomonitoringstat/AirAutoMonitoringStatDTO.java
@@ -23,8 +23,8 @@ public class AirAutoMonitoringStatDTO {
 
     Integer valError;
 
-    BigDecimal ratioReceivedDesign;
+    Double ratioReceivedDesign;
 
-    BigDecimal ratioErrorReceived;
+    Double ratioErrorReceived;
 
 }
