+++ src/main/java/com/EIPplatform/model/dto/report/airemmissionmanagement/airautoqcvnexceedance/AirAutoQcvnExceedanceDTO.java
@@ -21,5 +21,5 @@ public class AirAutoQcvnExceedanceDTO {
 
     Integer qcvnLimitValue;
 
-    BigDecimal exceedRatioPercent;
+    Double exceedRatioPercent;
 }
