+++ src/main/java/com/EIPplatform/model/dto/report/wastemanagement/domesticsolidwastestat/DomesticSolidWasteStatDTO.java
@@ -14,7 +14,7 @@ import java.time.LocalDateTime;
 public class DomesticSolidWasteStatDTO {
     Long domesticId;
     String wasteTypeName;
-    BigDecimal volumeCy;
+    Double volumeCy;
     String receiverOrg;
-    BigDecimal volumePy;
+    Double volumePy;
 }
