+++ src/main/java/com/EIPplatform/model/dto/report/wastemanagement/popinventorystat/PopInventoryStatDTO.java
@@ -17,9 +17,9 @@ public class PopInventoryStatDTO {
     String popName;
     String casCode;
     LocalDate importDate;
-    BigDecimal importVolume;
+    Double importVolume;
     String concentration;
-    BigDecimal volumeUsed;
-    BigDecimal volumeStocked;
+    Double volumeUsed;
+    Double volumeStocked;
     String complianceResult;
 }
