+++ src/main/java/com/EIPplatform/model/dto/report/wastemanagement/recycleindustrialwastestat/RecycleIndustrialWasteStatDTO.java
@@ -14,7 +14,7 @@ import java.time.LocalDateTime;
 public class RecycleIndustrialWasteStatDTO {
     Long recycleId;
     String transferOrg;
-    BigDecimal volumeCy;
+    Double volumeCy;
     String wasteTypeDesc;
-    BigDecimal volumePy;
+    Double volumePy;
 }
