+++ src/main/java/com/EIPplatform/model/dto/report/wastemanagement/selftreatedhwstat/SelfTreatedHwStatDTO.java
@@ -18,6 +18,6 @@ public class SelfTreatedHwStatDTO {
     Long selfTreatedId;
     String wasteName;
     String hwCode;
-    BigDecimal volumeKg;
+    Double volumeKg;
     String selfTreatmentMethod;
 }
